Make controller return only json
