+++ config/routes.rb
@@ -1,6 +1,11 @@
 Rails.application.routes.draw do
-  resources :skills
-  resources :users
+
+  namespace :api, { format: 'json' } do
+    namespace :v1 do
+      resources :skills
+      resources :users
+    end
+  end
   # The priority is based upon order of creation: first created -> highest priority.
   # See how all your routes lay out with "rake routes".
